renames controller package
